Fix bugs; add example project Used std.testing.refAllDeclsRecursive to find function signature errors   in the bindings Commented out timer.createTimerEx since I don't quite understand how it   should work, and C vaLists in Zig are broken on Windows, anyway. Added an example project in /example Updated libTES3MP-core to latest available commit Advanced interfaces (actor.zig, book.zig, etc) are still not ready, but   are now at least versions that reference the new binding layout.
TODO: figure out how to properly link a library, instead of the current
  BS with addLibraryPath and LinkSystemLibrary, which gets really messy

+++ src/binding/faction.zig
@@ -34,7 +34,7 @@ pub fn getFactionExpulsionState(pid: u16, index: u32) bool {
 
     return raw.getFactionExpulsionState(pid, index);
 }
-pub fn getFactionReputation(pid: u32, index: u32) i32 {
+pub fn getFactionReputation(pid: u16, index: u32) i32 {
     shared.triggerSafetyCheck(getFactionChangesSize(pid), index);
 
     return raw.getFactionReputation(pid, index);
